#31 Removes old pointer logic, renames pointer2.js -> pointer.js
